refactor: Refatorado o restante dos widgets para o padrão de arquitetura Atomic
